+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo1/TaskB.java
@@ -1,4 +1,4 @@
-package org.ohx.studyjuc.timer.timertest1;
+package org.ohx.studyjuc.timer.demo1;
 
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
