+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo1/Test3.java
@@ -1,25 +1,25 @@
-package org.ohx.studyjuc.timer.timertest1;
+package org.ohx.studyjuc.timer.demo1;
 
 import java.util.Date;
 import java.util.Timer;
 
 /**
- * 基本使用，定时任务线程不销毁
+ * 计划时间大于当前时间，定时任务立即执行
  *
  * @author mudkip
  * @date 2022/7/4
  */
-public class Test1 {
-    public static void main(String[] args) throws InterruptedException {
+public class Test3 {
+    public static void main(String[] args) {
         long start = System.currentTimeMillis();
         System.out.println("当前时间为：" + start);
-        long scheduleTime = start + 10000;
+        // 令计划时间大于当前时间
+        long scheduleTime = start - 5000;
         System.out.println("计划时间为：" + scheduleTime);
 
         TaskA task = new TaskA();
 
         Timer timer = new Timer();
-        Thread.sleep(1000);
         timer.schedule(task, new Date(scheduleTime));
     }
 }
