+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run1.java
@@ -1,4 +1,4 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Date;
 import java.util.Timer;
@@ -6,27 +6,26 @@ import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
- * schedule方法 任务延时
+ * schedule方法 任务不延时
  *
  * @author mudkip
  * @date 2022/7/4
  */
-public class Run3 {
+public class Run1 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
-        System.out.println("now time " + System.currentTimeMillis());
         Timer timer = new Timer();
         timer.schedule(new TimerTask() {
             @Override
             public void run() {
                 try {
                     System.out.println("begin timer = " + System.currentTimeMillis());
-                    TimeUnit.SECONDS.sleep(5);
+                    TimeUnit.SECONDS.sleep(1);
                     System.out.println("end timer = " + System.currentTimeMillis());
                 } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
-        }, new Date(now), 2000);
+        }, new Date(now), 3000);
     }
 }
