+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run10.java
@@ -1,16 +1,16 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Date;
 import java.util.Timer;
 import java.util.TimerTask;
 
 /**
- * schedule方法不具有追赶性
+ * scheduleAtFixedRate方法具有追赶性
  *
  * @author mudkip
  * @date 2022/7/5
  */
-public class Run9 {
+public class Run10 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
@@ -19,7 +19,7 @@ public class Run9 {
         System.out.println("schedule time " + System.currentTimeMillis());
 
         Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
+        timer.scheduleAtFixedRate(new TimerTask() {
             @Override
             public void run() {
                 System.out.println("begin timer = " + System.currentTimeMillis());
