+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run2.java
@@ -1,31 +1,31 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
- * scheduleAtFixedRate方法 延时
+ * schedule方法 任务不延时
  *
  * @author mudkip
- * @date 2022/7/5
+ * @date 2022/7/4
  */
-public class Run8 {
+public class Run2 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
         Timer timer = new Timer();
-        timer.scheduleAtFixedRate(new TimerTask() {
+        timer.schedule(new TimerTask() {
             @Override
             public void run() {
                 try {
                     System.out.println("begin timer = " + System.currentTimeMillis());
-                    TimeUnit.SECONDS.sleep(5);
+                    TimeUnit.SECONDS.sleep(1);
                     System.out.println("end timer = " + System.currentTimeMillis());
                 } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
-        }, 3000, 2000);
+        }, 3000, 4000);
     }
 }
