+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run3.java
@@ -1,4 +1,4 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Date;
 import java.util.Timer;
@@ -6,17 +6,17 @@ import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
- * scheduleAtFixedRate方法 延时
+ * schedule方法 任务延时
  *
  * @author mudkip
- * @date 2022/7/5
+ * @date 2022/7/4
  */
-public class Run7 {
+public class Run3 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
         Timer timer = new Timer();
-        timer.scheduleAtFixedRate(new TimerTask() {
+        timer.schedule(new TimerTask() {
             @Override
             public void run() {
                 try {
