+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run6.java
@@ -1,21 +1,20 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
- * schedule方法 任务不延时
+ * scheduleAtFixedRate方法 不延时
  *
  * @author mudkip
- * @date 2022/7/4
+ * @date 2022/7/5
  */
-public class Run2 {
+public class Run6 {
     public static void main(String[] args) {
-        long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
         Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
+        timer.scheduleAtFixedRate(new TimerTask() {
             @Override
             public void run() {
                 try {
