+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run8.java
@@ -1,17 +1,16 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
-import java.util.Date;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 /**
- * scheduleAtFixedRate方法 不延时
+ * scheduleAtFixedRate方法 延时
  *
  * @author mudkip
  * @date 2022/7/5
  */
-public class Run5 {
+public class Run8 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
@@ -21,12 +20,12 @@ public class Run5 {
             public void run() {
                 try {
                     System.out.println("begin timer = " + System.currentTimeMillis());
-                    TimeUnit.SECONDS.sleep(1);
+                    TimeUnit.SECONDS.sleep(5);
                     System.out.println("end timer = " + System.currentTimeMillis());
                 } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
-        }, new Date(now), 3000);
+        }, 3000, 2000);
     }
 }
