+++ study-juc/src/main/java/org/ohx/studyjuc/timer/demo2/Run9.java
@@ -1,16 +1,16 @@
-package org.ohx.studyjuc.timer.timertest2;
+package org.ohx.studyjuc.timer.demo2;
 
 import java.util.Date;
 import java.util.Timer;
 import java.util.TimerTask;
 
 /**
- * scheduleAtFixedRate方法具有追赶性
+ * schedule方法不具有追赶性
  *
  * @author mudkip
  * @date 2022/7/5
  */
-public class Run10 {
+public class Run9 {
     public static void main(String[] args) {
         long now = System.currentTimeMillis();
         System.out.println("now time " + System.currentTimeMillis());
@@ -19,7 +19,7 @@ public class Run10 {
         System.out.println("schedule time " + System.currentTimeMillis());
 
         Timer timer = new Timer();
-        timer.scheduleAtFixedRate(new TimerTask() {
+        timer.schedule(new TimerTask() {
             @Override
             public void run() {
                 System.out.println("begin timer = " + System.currentTimeMillis());
