--- study-juc/src/main/java/org/ohx/studyjuc/timer/timertest1/README.md
@@ -1,1 +1,0 @@
-Timer 和 TimerTask 的基本使用，错误示例与正确示例。
