--- study-juc/src/main/java/org/ohx/studyjuc/timer/timertest2/README.md
@@ -1,10 +1,0 @@
-schedule() 方法和 scheduleAtFixedRate() 方法的主要区别在于有没有追赶特性：
-当定时任务实际启动时间晚于计划启动时间时，这段丢失的时间内应该执行的任务，schedule() 是不会执行的；而 scheduleAtFixedRate() 则会补上。
-
-对于 schedule() 方法：
-* 当任务没有被延时执行，则下一次执行开始的时间是上一次任务的开始时间加上period时间（“不延时”是指执行任务的时间小于period间隔时间。）。 
-* 当任务被延时执行，则下一次任务的执行时间参考的是上一次任务“结束”时的时间来开始（“延时”是指执行任务的时间大于period间隔时间。）。
-
-对于 scheduleAtFixedRate() 方法：
-* 当任务没有被延时执行，则下一次执行开始的时间是上一次任务的开始时间加上period时间（“不延时”是指执行任务的时间小于period间隔时间。）。 
-* 当任务被延时执行，则下一次任务的执行时间参考的是上一次任务“结束”时的时间来开始（“延时”是指执行任务的时间大于period间隔时间。）。
